+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+* Attribute dictionary added to class `compas_rcf.fabrication.clay_obj.ClayBullet`
+
+### Changed
+* Property `Vector` in `ClayBullet` changed to represent the bullets center as a line.
+
+### Removed
+
 ## \[0.1.14\] \[2020-10-14\]
 
 ### Added